Fix platform-dependent types: change file_index from usize to u64
- Changed file_index type from usize to u64 in InodeEntry::File variant
- Updated all internal usages in filesystem.rs, async_bridge.rs, and inode.rs
- Added explicit casts at API boundaries where rqbit expects usize
- Ensures 32-bit and 64-bit platform compatibility for file indices
- All tests pass, clippy clean

+++ src/fs/async_bridge.rs
@@ -16,7 +16,7 @@ pub enum FuseRequest {
         /// Torrent ID
         torrent_id: u64,
         /// File index within the torrent
-        file_index: usize,
+        file_index: u64,
         /// Offset to start reading from
         offset: u64,
         /// Number of bytes to read
@@ -171,7 +171,7 @@ impl AsyncFuseWorker {
 
                 let result = tokio::time::timeout(
                     timeout,
-                    api_client.read_file_streaming(torrent_id, file_index, offset, size),
+                    api_client.read_file_streaming(torrent_id, file_index as usize, offset, size),
                 )
                 .await;
 
@@ -290,7 +290,7 @@ impl AsyncFuseWorker {
     pub fn read_file(
         &self,
         torrent_id: u64,
-        file_index: usize,
+        file_index: u64,
         offset: u64,
         size: usize,
         timeout: Duration,
